Compute cpu relative to last sample
ps is great, but the cpu % value it spouts is relative to when the
process started. If jackett has been running a long time, it can take
quite a long time for that % value to be affected to the point that we
take action.

This version looks at cpu time at each loop and computes the % cpu since
the last time it looked rather than the start of the process.

It's still rather experimental: there are no command line options, no
documentation, and it's hardwired to look only at Jackett.